requirements.txt
streamlit
langchain
langchain-google-genai
langchain-chroma
langchain-core
pypdf
python-dotenv
chromadb
langchain-community
beautifulsoup4